Use single file for tests
